Update pline_offset_scene.rs
On my local machine, clippy complains about an unneeded strut pattern. Removing the struct pattern causes it to not compile, so disabling the check with [allow(clippy::unneeded_struct_pattern)] is required. However, apparently the version of clippy on github does not support this option, and that caused a CI failure.

This update removes the [allow(clippy::unneeded_struct_pattern)].